Tirando import e adicionando outro

+++ src/pages/Produto/index.jsx
@@ -1,16 +1,16 @@
 import { useEffect, useState } from 'react';
 import { useParams, Link, useNavigate } from 'react-router-dom';
-import { Banner } from '../../componentes/Banner/index.jsx';
+import { jwtDecode } from 'jwt-decode';
 import { BarraRodape } from '../../componentes/BarraRodape/index.jsx';
 import { DetalheProduto } from '../../componentes/DetalheProduto/index.jsx';
 import { Botao } from '../../componentes/Botao/index.jsx';
-import api from '../../services/api.js'; 
+import api from '../../services/api.js';
 import { NavBar } from '../../componentes/NavBar/index.jsx';
 
 export const Produto = () => {
     const { id } = useParams();
     const navigate = useNavigate();
-    const [produto, setProduto] = useState(null); 
+    const [produto, setProduto] = useState(null);
 
     useEffect(() => {
         const carregarProduto = async () => {
@@ -20,18 +20,32 @@ export const Produto = () => {
             } catch (erro) {
                 console.error("Erro ao buscar produto:", erro);
                 alert("Erro ao carregar o produto!");
-                navigate('/'); 
+                navigate('/');
             }
         };
         carregarProduto();
     }, [id, navigate]);
 
     const excluirProduto = async () => {
+        const token = localStorage.getItem('token');
+        let usuarioLogadoId = null;
+
+        if (token) {
+            const decoded = jwtDecode(token);
+            usuarioLogadoId = decoded.id;
+        }
+
+        if (String(usuarioLogadoId) !== String(id)) {
+            alert("Você não pode excluir um produto de outro usuário!")
+            navigate(`/produto/${id}`)
+            return;
+        }
+
         if (window.confirm("Tem certeza que deseja excluir este produto?")) {
             try {
                 await api.delete(`/shop/${id}`);
                 alert("Produto excluído com sucesso!");
-                navigate('/produtos'); 
+                navigate('/produtos');
             } catch (erro) {
                 console.error("Erro ao excluir:", erro);
                 alert("Não foi possível excluir o produto.");
@@ -50,21 +64,42 @@ export const Produto = () => {
         );
     }
 
+    const verificarEdição = async () => {
+        const token = localStorage.getItem('token');
+        let usuarioLogadoId = null;
+
+        if (token) {
+            const decoded = jwtDecode(token);
+            usuarioLogadoId = decoded.id;
+        }
+
+        const idVendedor = produto.soldBy
+
+        if (String(usuarioLogadoId) !== String(idVendedor)) {
+            alert("Você não pode editar o produto de outro usuário!")
+            navigate(`/produto/${id}`)
+            return;
+        }
+
+        navigate("/cadastro-produto", { state: { produtoParaEditar: produto } })
+    }
+
+
     return (
         <>
-            <NavBar/>
-            <DetalheProduto 
-                nome={produto.name} 
-                descricao={produto.description} 
-                preco={`R$ ${produto.price}`} 
+            <NavBar />
+            <DetalheProduto
+                nome={produto.name}
+                descricao={produto.description}
+                preco={`R$ ${produto.price}`}
                 img={produto.imageUrl || 'https://github.com/kawevk.png'}
             />
-            
+
             <div style={{ display: 'flex', justifyContent: 'center', gap: '20px' }}>
-                <Link to={`/cadastro-produto`} state={{ produtoParaEditar: produto }}>
+                <div onClick={verificarEdição}>
                     <Botao className={'botao-padrao'}>Editar Produto</Botao>
-                </Link>
-                
+                </div>
+
                 <div onClick={excluirProduto}>
                     <Botao className={'botao-excluir'}>Excluir Produto</Botao>
                 </div>
